refactor(e2e): modal-navigation app tests
