+++ e2e/modal-navigation/e2e/e2e-spec-app-root/app-root-modal-frame-background.e2e-spec.ts
@@ -4,7 +4,7 @@ import { modalFrameBackground,
     testSecondPageBackground,
     testNestedModalFrameBackground,
     testNestedModalPageBackground
-} from "../e2e-spec-shared/modal-frame"
+} from "../shared.e2e-spec"
 
 describe("app root modal frame background scenarios", () => {
