+++ e2e/modal-navigation/e2e/e2e-spec-tab-root/tab-root-modal-page-background.e2e-spec.ts
@@ -4,7 +4,7 @@ import { modalPageBackground,
     testSecondPageBackground,
     testNestedModalFrameBackground,
     testNestedModalPageBackground
-} from "../e2e-spec-shared/modal-frame"
+} from "../shared.e2e-spec"
 
 describe("tab root modal page background scenarios", () => {
